+++ exercises/concept/tim-from-marketing/.docs/instructions.md
@@ -4,7 +4,8 @@ In this exercise you'll be writing code to print name badges for factory employe
 
 ## 1. Print a badge for an employee
 
-Employees have an ID, name and department name. Employee badge labels are formatted as follows: `"[id] - name - DEPARTMENT"`. Implement the (_static_) `Badge.print()` method to return an employee's badge label:
+Employees have an ID, name and department name. Employee badge labels are formatted as follows: `"[id] - name - DEPARTMENT"`.
+Implement the `Badge.print()` method to return an employee's badge label:
 
 ```java
 Badge badge = new Badge();
@@ -16,21 +17,24 @@ Note that the department should be uppercased on the label.
 
 ## 2. Print a badge for a new employee
 
-Due to a quirk in the computer system, new employees occasionally don't yet have an ID when they start working at the factory. As badges are required, they will receive a temporary badge without the ID prefix. Modify the (_static_) `Badge.print()` method to support new employees that don't yet have an ID:
+Due to a quirk in the computer system, new employees occasionally don't yet have an ID when they start working at the factory.
+As badges are required, they will receive a temporary badge without the ID prefix. Modify the `Badge.print()` method to support new employees that don't yet have an ID:
 
 ```java
 Badge badge = new Badge();
-Badge.print(id: null, "Jane Johnson", "Procurement");
+Badge.print(null, "Jane Johnson", "Procurement");
 // => "Jane Johnson - PROCUREMENT"
 ```
 
 ## 3. Print a badge for the owner
 
-Even the factory's owner has to wear a badge at all times. However, an owner does not have a department. In this case, the label should print `"OWNER"` instead of the department name. Modify the (_static_) `Badge.print()` method to print a label for the owner:
+Even the factory's owner has to wear a badge at all times.
+However, an owner does not have a department. In this case, the label should print `"OWNER"` instead of the department name.
+Modify the `Badge.print()` method to print a label for the owner:
 
 ```java
 Badge badge = new Badge();
-badge.print(254, "Charlotte Hale", department: null);
+badge.print(254, "Charlotte Hale", null);
 // => "[254] - Charlotte Hale - OWNER"
 ```
 
@@ -38,6 +42,6 @@ Note that it is possible for the owner to also be a new employee:
 
 ```java
 Badge badge = new Badge();
-badge.print(id: null, "Charlotte Hale", department: null);
+badge.print(null, "Charlotte Hale", null);
 // => "Charlotte Hale - OWNER"
 ```
