+++ exercises/concept/tim-from-marketing/.docs/introduction.md
@@ -4,7 +4,8 @@
 
 In Java, the `null` literal is used to denote the absence of a value.
 
-Primitive data types in java all have a default value and therefore can never be `null`. By convention they start with a lowercase letter e.g `int`
+Primitive data types in java all have a default value and therefore can never be `null`.
+By convention they start with a lowercase letter e.g `int`
 
 Reference types contain the memory address of an object can
 have a value of null. These variables usually start with an uppercase e.g `String`
