+++ exercises/concept/tim-from-marketing/src/main/java/Badge.java
@@ -1,5 +1,5 @@
-public class Badge {
-    public static String print(Integer id, String name, String department) {
+class Badge {
+    public String print(Integer id, String name, String department) {
         throw new UnsupportedOperationException("Please implement the (static) Badge.print() method");
     }
 }
